contact liss class

+++ src/Contact.java
@@ -2,9 +2,9 @@ public class Contact {
 
     String firstName;
     String lastName;
-    Long phoneNumber;
+    int phoneNumber;
 
-    public Contact (String firstName, String lastName, Long phoneNumber) {
+    public Contact (String firstName, String lastName, int phoneNumber) {
         this.firstName = firstName;
         this.lastName = lastName;
         this.phoneNumber = phoneNumber;
